+++ docs/changelog.rst
@@ -1,5 +1,6 @@
 0.6.0
 -----
+- deprecated `alter_column` comparing of columns. Just use explicit parameter change.
 - added option to define custom templates through option ``--templates_path`` and ``--templates_theme``, read more in :ref:`tutorial section <custom-templates>`
 - use Python logging for output, can be shut down by passing ``--disable_logging`` to :func:`migrate.versioning.shell.main`
 - `url` parameter can also be an :class:`Engine` instance (this usage is discouraged though sometimes necessary)